i moved the code from PlayerStats to CharacterStatsHandler. Deleted PlayersStats
