--- Assets/Scripts/DamageOnTouch.cs
@@ -1,24 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class DamageOnTouch : MonoBehaviour
-{
-    public int damageAmount = 10;
-    private PlayerStats playerStats;
-
-    void Start()
-    {
-
-        playerStats = GameObject.FindGameObjectWithTag("Player").GetComponent<PlayerStats>();
-    }
-
-    private void OnTriggerEnter2D(Collider2D other)
-    {
-        if (other.CompareTag("Player"))
-        {
-            playerStats.TakeDamage(damageAmount);
-        }
-    }
-}
-
