--- Assets/Scripts/PlayerStats.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 17dc149b6a3f79744b9b41a1f13ac6e7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
